backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
werkzeug==3.0.1
cryptography==41.0.7
python-dateutil==2.8.2
email-validator==2.1.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
Faker==22.2.0
flasgger==0.9.7.1
flask-apispec==0.11.4
webargs==8.4.0
psycopg2-binary==2.9.9
